refactor: extract worker
